backend/fastapi_keycloak/api/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
async-property==0.2.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
deprecation==2.1.0
fastapi==0.112.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jwcrypto==1.5.6
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
PyJWT==2.9.0
python-dotenv==1.0.1
python-keycloak==4.3.0
python-multipart==0.0.9
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==72.1.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
wheel==0.43.0
